Move RepositoryFullName from GitHub to just SourceControl
