+++ Source/Core.Specs/SourceControl/GitHub/for_InstallationsWebhookHandler/when_handling_installation_repositories_event/and_there_are_repository_additions_and_removals.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using Concepts.SourceControl;
 using Concepts.SourceControl.GitHub;
 using Dolittle.Commands;
 using Domain.SourceControl.GitHub;
