+++ Source/Core/SourceControl/GitHub/InstallationsWebhookHandler.cs
@@ -5,6 +5,7 @@
 
 using System;
 using System.Linq;
+using Concepts.SourceControl;
 using Concepts.SourceControl.GitHub;
 using Dolittle.Commands;
 using Dolittle.Commands.Coordination;
