+++ Source/Domain/SourceControl/GitHub/Application.cs
@@ -6,6 +6,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Concepts.SourceControl;
 using Concepts.SourceControl.GitHub;
 using Dolittle.Domain;
 using Dolittle.Runtime.Events;
