+++ Source/Domain/SourceControl/GitHub/ApplicationCommandHandler.cs
@@ -3,6 +3,7 @@
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  * --------------------------------------------------------------------------------------------*/
 
+using Concepts.SourceControl;
 using Concepts.SourceControl.GitHub;
 using Dolittle.Collections;
 using Dolittle.Commands;
